Move non-CLI code into separate library
