+++ NorModifierLib/Exceptions/BiosReadException.cs
@@ -1,3 +1,3 @@
-﻿namespace UART_CL_By_TheCod3r.Exceptions;
+﻿namespace NorModifierLib.Exceptions;
 
 public class BiosReadException(string message, Exception innerException) : Exception(message, innerException) { }
