+++ NorModifierLib/Exceptions/BiosWriteException.cs
@@ -1,3 +1,3 @@
-﻿namespace UART_CL_By_TheCod3r.Exceptions;
+﻿namespace NorModifierLib.Exceptions;
 
 public class BiosWriteException(string message, Exception innerException) : Exception(message, innerException) { }
