+++ PS5 NOR Modifier.sln
@@ -9,6 +9,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "UART-CL By TheCod3r", "UART
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "UART-CL Tests", "UART-CL Tests\UART-CL Tests.csproj", "{0E155073-52EB-4771-A9CB-E6FD7DC42D2E}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NorModifierLib", "NorModifierLib\NorModifierLib.csproj", "{5E047A92-075B-4C7F-87CA-B708320F80A9}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -41,6 +43,14 @@ Global
 		{0E155073-52EB-4771-A9CB-E6FD7DC42D2E}.Release|Any CPU.Build.0 = Release|Any CPU
 		{0E155073-52EB-4771-A9CB-E6FD7DC42D2E}.Release|x86.ActiveCfg = Release|Any CPU
 		{0E155073-52EB-4771-A9CB-E6FD7DC42D2E}.Release|x86.Build.0 = Release|Any CPU
+		{5E047A92-075B-4C7F-87CA-B708320F80A9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5E047A92-075B-4C7F-87CA-B708320F80A9}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5E047A92-075B-4C7F-87CA-B708320F80A9}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{5E047A92-075B-4C7F-87CA-B708320F80A9}.Debug|x86.Build.0 = Debug|Any CPU
+		{5E047A92-075B-4C7F-87CA-B708320F80A9}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5E047A92-075B-4C7F-87CA-B708320F80A9}.Release|Any CPU.Build.0 = Release|Any CPU
+		{5E047A92-075B-4C7F-87CA-B708320F80A9}.Release|x86.ActiveCfg = Release|Any CPU
+		{5E047A92-075B-4C7F-87CA-B708320F80A9}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
