+++ UART-CL By TheCod3r/UART-CL By TheCod3r.csproj
@@ -37,4 +37,8 @@
     <PackageReference Include="System.IO.Ports" Version="9.0.5" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\NorModifierLib\NorModifierLib.csproj" />
+  </ItemGroup>
+
 </Project>
